Rename ttnn/api/ttnn/metal2_artifacts.hpp -> metal_v2_artifacts.hpp
Match the metal_v2 naming. Updates the two #include sites and the
ttnn/sources.cmake public-API header entry. File content unchanged.
